+++ src/index.js
@@ -18,10 +18,7 @@ module.exports = function (config) {
 
   if (!config) {
     // CLI
-    config = appendOptions(
-      program.version(version).allowUnknownOption(true),
-      options,
-    )
+    config = appendOptions(program.version(version).allowUnknownOption(true), options)
       .parse(process.argv)
       .opts();
   }
@@ -62,8 +59,7 @@ module.exports = function (config) {
   }
 
   if (config.autoRelay) {
-    const emailAddress =
-      typeof config.autoRelay === "string" ? config.autoRelay : null;
+    const emailAddress = typeof config.autoRelay === "string" ? config.autoRelay : null;
     mailserver.setAutoRelayMode(true, config.autoRelayRules, emailAddress);
   }